+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Initial release of @react-awesome-query-builder/radix-ui
+- Initial release of react-awesome-query-builder-radix-ui
 - Core widgets:
   - RadixButton - Action buttons with primary/danger variants
   - RadixButtonGroup - Button grouping component